+++ src/main/java/com/uspray/uspray/controller/PrayController.java
@@ -1,9 +1,9 @@
 package com.uspray.uspray.controller;
 
 
-import com.uspray.uspray.common.dto.ApiResponseDto;
-import com.uspray.uspray.domain.dto.PrayDto;
-import com.uspray.uspray.domain.dto.request.PrayRequestDto;
+import com.uspray.uspray.dto.ApiResponseDto;
+import com.uspray.uspray.dto.pray.PrayDto;
+import com.uspray.uspray.dto.pray.request.PrayRequestDto;
 import com.uspray.uspray.exception.SuccessStatus;
 import com.uspray.uspray.service.PrayService;
 import io.swagger.v3.oas.annotations.Operation;
