+++ src/main/java/com/uspray/uspray/dto/pray/PrayDto.java
@@ -1,4 +1,4 @@
-package com.uspray.uspray.domain.dto;
+package com.uspray.uspray.dto.pray;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
